feat: split post page in a different component
